Extract refactor the harvest stuff
Better names. API was a dumping ground
